fix: proper error surfacing for permission errors during scope expansion
