+++ rbac/adapter/permission.go
@@ -13,9 +13,7 @@ import (
 	"github.com/flanksource/commons/collections"
 	"github.com/flanksource/duty/context"
 	"github.com/flanksource/duty/models"
-	dutyRBAC "github.com/flanksource/duty/rbac"
 	pkgPolicy "github.com/flanksource/duty/rbac/policy"
-	"github.com/flanksource/duty/types"
 	gocache "github.com/patrickmn/go-cache"
 	"github.com/samber/lo"
 	"gorm.io/gorm"
@@ -50,7 +48,7 @@ func (a *PermissionAdapter) LoadPolicy(model model.Model) error {
 	}
 
 	var permissions []models.Permission
-	if err := a.db.Where("deleted_at IS NULL").Find(&permissions).Error; err != nil {
+	if err := a.db.Where("deleted_at IS NULL AND error IS NULL").Find(&permissions).Error; err != nil {
 		return fmt.Errorf("failed to load permissions: %w", err)
 	}
 
@@ -58,6 +56,16 @@ func (a *PermissionAdapter) LoadPolicy(model model.Model) error {
 		// Expand scope references in object_selector before converting to Casbin rules
 		expandedPerm, err := a.expandPermissionScopes(permission)
 		if err != nil {
+			var validationErr *scopeExpansionValidationError
+			if errors.As(err, &validationErr) {
+				// Persist validation error to database
+				if updateErr := a.db.Model(&permission).Update("error", err.Error()).Error; updateErr != nil {
+					return fmt.Errorf("failed to update permission error: %w", updateErr)
+				}
+
+				continue // Skip this permission
+			}
+
 			return err
 		}
 
@@ -263,101 +271,4 @@ func (a *PermissionAdapter) permissionGroupToCasbinRule(permission models.Permis
 	}
 
 	return policies, nil
-}
-
-// getScopeTargets retrieves scope targets by namespace and name, using cache when available.
-// Returns gorm.ErrRecordNotFound if scope doesn't exist.
-// Returns json unmarshal error if scope.Targets contains invalid JSON.
-func (a *PermissionAdapter) getScopeTargets(namespace, name string) ([]v1.ScopeTarget, error) {
-	cacheKey := namespace + "/" + name
-
-	if cached, found := a.cache.Get(cacheKey); found {
-		return cached.([]v1.ScopeTarget), nil
-	}
-
-	var scope models.Scope
-	err := a.db.
-		Where("name = ? AND namespace = ? AND deleted_at IS NULL", name, namespace).
-		First(&scope).Error
-	if err != nil {
-		return nil, err
-	}
-
-	var targets []v1.ScopeTarget
-	if err := json.Unmarshal([]byte(scope.Targets), &targets); err != nil {
-		return nil, fmt.Errorf("failed to unmarshal scope targets for %s/%s: %w", namespace, name, err)
-	}
-
-	a.cache.Set(cacheKey, targets, gocache.DefaultExpiration)
-
-	return targets, nil
-}
-
-// expandPermissionScopes expands scope references in a permission's object_selector
-// and returns a new permission with the expanded selectors merged in.
-func (a *PermissionAdapter) expandPermissionScopes(perm models.Permission) (models.Permission, error) {
-	// If no object selector, nothing to expand
-	if len(perm.ObjectSelector) == 0 {
-		return perm, nil
-	}
-
-	var selectors dutyRBAC.Selectors
-	if err := json.Unmarshal(perm.ObjectSelector, &selectors); err != nil {
-		return perm, fmt.Errorf("failed to unmarshal object_selector: %w", err)
-	}
-
-	// If no scope references, return as-is
-	if len(selectors.Scopes) == 0 {
-		return perm, nil
-	}
-
-	// Expand scopes and merge into selectors
-	for _, scopeRef := range selectors.Scopes {
-		targets, err := a.getScopeTargets(scopeRef.Namespace, scopeRef.Name)
-		if err != nil {
-			if errors.Is(err, gorm.ErrRecordNotFound) {
-				// Scope not found - skip
-				continue
-			}
-
-			return perm, err
-		}
-
-		// Merge targets into selectors (union approach)
-		for _, target := range targets {
-			if target.Config != nil {
-				selectors.Configs = append(selectors.Configs, convertScopeResourceSelectorToResourceSelector(target.Config))
-			}
-			if target.Component != nil {
-				selectors.Components = append(selectors.Components, convertScopeResourceSelectorToResourceSelector(target.Component))
-			}
-			if target.Playbook != nil {
-				selectors.Playbooks = append(selectors.Playbooks, convertScopeResourceSelectorToResourceSelector(target.Playbook))
-			}
-			// Note: Canary targets are skipped - dutyRBAC.Selectors doesn't have a Canaries field
-			// Note: Global targets are ignored per design decision
-			// Note: Connection targets are ignored (no RLS support yet per auth/rls.go:127-132)
-		}
-	}
-
-	// Marshal the expanded selectors back to JSON
-	expandedObjectSelector, err := json.Marshal(selectors)
-	if err != nil {
-		return perm, fmt.Errorf("failed to marshal expanded selectors: %w", err)
-	}
-
-	// Create a new permission with the expanded object selector
-	expandedPerm := perm
-	expandedPerm.ObjectSelector = expandedObjectSelector
-
-	return expandedPerm, nil
-}
-
-// convertScopeResourceSelectorToResourceSelector converts a v1.ScopeResourceSelector to types.ResourceSelector
-func convertScopeResourceSelectorToResourceSelector(scopeSel *v1.ScopeResourceSelector) types.ResourceSelector {
-	return types.ResourceSelector{
-		Agent:       scopeSel.Agent,
-		Name:        scopeSel.Name,
-		TagSelector: scopeSel.TagSelector,
-	}
 }
